validation: Better more consistent validation during hostname checks
We now use the validators library for validation checks which allows for
more consistent validation.  This is done via the new is_valid()
function which rids the need for two seperate functions.  If the check
succeeds for both IPv4 or FQDN we consider that a success and try to
authenticate via ssh next.

Signed-off-by: Kyle Squizzato <ksquizz@gmail.com>